test(workflows): stop on fail

+++ test/Neuroglia.UnitTests/Cases/Data/Repositories/DistributedCacheRepositoryTests.cs
@@ -39,6 +39,7 @@ namespace Neuroglia.UnitTests.Cases.Data.Repositories
         [Fact, Priority(0)]
         public async Task AddEntity()
         {
+            if (StopOnFail.HasFail) return;
             //arrange
             var firstName = "Fake First Name";
             var lastName = "Fake Last Name";
@@ -62,6 +63,7 @@ namespace Neuroglia.UnitTests.Cases.Data.Repositories
         [Fact, Priority(1)]
         public async Task ContainsEntity()
         {
+            if (StopOnFail.HasFail) return;
             //act
             bool exists = await this.Repository.ContainsAsync(EntityId.Value);
 
@@ -72,6 +74,7 @@ namespace Neuroglia.UnitTests.Cases.Data.Repositories
         [Fact, Priority(2)]
         public async Task FindEntity()
         {
+            if (StopOnFail.HasFail) return;
             //act
             TestPerson entity = await this.Repository.FindAsync(EntityId.Value);
 
@@ -87,6 +90,7 @@ namespace Neuroglia.UnitTests.Cases.Data.Repositories
         [Fact, Priority(3)]
         public async Task UpdateEntity()
         {
+            if (StopOnFail.HasFail) return;
             //arrange
             var entity = await this.Repository.FindAsync(EntityId.Value);
             var newFirstName = "Updated Fake First Name";
@@ -110,6 +114,7 @@ namespace Neuroglia.UnitTests.Cases.Data.Repositories
         [Fact, Priority(4)]
         public async Task ListEntities()
         {
+            if (StopOnFail.HasFail) return;
             //act
             var entities = await this.Repository.ToListAsync();
 
@@ -120,6 +125,7 @@ namespace Neuroglia.UnitTests.Cases.Data.Repositories
         [Fact, Priority(5)]
         public void GetQueryable()
         {
+            if (StopOnFail.HasFail) return;
             //act
             var entities = this.Repository.AsQueryable().ToList();
 
@@ -130,6 +136,7 @@ namespace Neuroglia.UnitTests.Cases.Data.Repositories
         [Fact, Priority(6)]
         public async Task RemoveEntity()
         {
+            if (StopOnFail.HasFail) return;
             //act
             await this.Repository.RemoveAsync(EntityId.Value);
             await this.Repository.SaveChangesAsync();
@@ -142,6 +149,8 @@ namespace Neuroglia.UnitTests.Cases.Data.Repositories
         [Fact, Priority(10)]
         public async Task End()
         {
+            if (StopOnFail.HasFail) return;
+            StopOnFail.HasFail = true;
             throw new NotImplementedException("DistributedCacheRepositoryTests");
         }
