+++ test/Neuroglia.UnitTests/Cases/Data/Repositories/EFCoreRepositoryTests.cs
@@ -40,6 +40,7 @@ namespace Neuroglia.UnitTests.Cases.Data.Repositories
         [Fact, Priority(0)]
         public async Task AddEntity()
         {
+            if (StopOnFail.HasFail) return;
             //arrange
             var firstName = "Fake First Name";
             var lastName = "Fake Last Name";
@@ -64,6 +65,7 @@ namespace Neuroglia.UnitTests.Cases.Data.Repositories
         [Fact, Priority(1)]
         public async Task ContainsEntity()
         {
+            if (StopOnFail.HasFail) return;
             //act
             bool exists = await this.Repository.ContainsAsync(EntityId.Value);
 
@@ -74,6 +76,7 @@ namespace Neuroglia.UnitTests.Cases.Data.Repositories
         [Fact, Priority(2)]
         public async Task FindEntity()
         {
+            if (StopOnFail.HasFail) return;
             //act
             TestPerson entity = await this.Repository.FindAsync(EntityId.Value);
 
@@ -85,6 +88,7 @@ namespace Neuroglia.UnitTests.Cases.Data.Repositories
         [Fact, Priority(3)]
         public async Task UpdateEntity()
         {
+            if (StopOnFail.HasFail) return;
             //arrange
             var entity = await this.Repository.FindAsync(EntityId.Value);
             var newFirstName = "Updated Fake First Name";
@@ -108,6 +112,7 @@ namespace Neuroglia.UnitTests.Cases.Data.Repositories
         [Fact, Priority(4)]
         public async Task ListEntities()
         {
+            if (StopOnFail.HasFail) return;
             //act
             var entities = await this.Repository.ToListAsync();
 
@@ -118,6 +123,7 @@ namespace Neuroglia.UnitTests.Cases.Data.Repositories
         [Fact, Priority(5)]
         public async Task GetQueryable()
         {
+            if (StopOnFail.HasFail) return;
             //act
             var entities = await this.Repository.AsQueryable().ToListAsync();
 
@@ -128,6 +134,7 @@ namespace Neuroglia.UnitTests.Cases.Data.Repositories
         [Fact, Priority(6)]
         public async Task RemoveEntity()
         {
+            if (StopOnFail.HasFail) return;
             //act
             await this.Repository.RemoveAsync(EntityId.Value);
             await this.Repository.SaveChangesAsync();
@@ -140,6 +147,8 @@ namespace Neuroglia.UnitTests.Cases.Data.Repositories
         [Fact, Priority(10)]
         public async Task End()
         {
+            if (StopOnFail.HasFail) return;
+            StopOnFail.HasFail = true;
             throw new NotImplementedException("EFCoreRepositoryTests");
         }
